+++ lib/gandi_v5/email.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'email/mailbox'
-require_relative 'email/offer'
-require_relative 'email/slot'
-
 class GandiV5
   # Gandi Email Mailbox Management API.
   class Email