+++ lib/gandi_v5/error.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'error/gandi_error'
-
 class GandiV5
   # Generic error class for errors occuring using the API.
   class Error < RuntimeError